refactor : step5 - MVC 패턴 기반으로 리팩토링 * Application의 시작인 클래스에서 모든 로직이 들어가는 구조를 변경 ** Controller, Service 분리 ** web application 이라고 생각하면 Service에서 DTO를 반환하면 좋지만, 해당 어플리케이션은 값을 출력하고 끝나므로 return은 없음
