+++ src/main/java/racing/RacingApplication.java
@@ -1,0 +1,10 @@
+package racing;
+
+import racing.controller.RacingController;
+
+public class RacingApplication {
+
+    public static void main(String[] args) {
+        RacingController.play();
+    }
+}
