+++ src/main/java/racing/service/RacingService.java
@@ -1,30 +1,18 @@
-package racing;
+package racing.service;
 
-import racing.domain.*;
-import racing.view.InputView;
+import racing.domain.Car;
+import racing.domain.Cars;
+import racing.domain.Movable;
 import racing.view.ResultView;
 
 import java.util.Random;
 
-public class RacingGame {
+public class RacingService {
     private static final int RANDOM_VALUE_MAX = 10;
-    public static final int MOVE_CONDITION = 4;
+    private static final int MOVE_CONDITION = 4;
     private static final String CARS_NAME_DELIMITER = ",";
 
-    private static String carNames;
-    private static int tryCnt;
-
-    private static void gameSetting() {
-        do {
-            carNames = InputView.getCarNames();
-        } while(!ValidChecker.checkInput(carNames));
-
-        do {
-            tryCnt = InputView.getTryCnt();
-        } while (!ValidChecker.checkInput(tryCnt));
-    }
-
-    private static void playGame() {
+    public void playGame(String carNames, int tryCnt) {
         Cars cars = new Cars(carNames.split(CARS_NAME_DELIMITER));
         StringBuilder stringBuilder = ResultView.getResultViewBuilder();
 
@@ -38,7 +26,7 @@ public class RacingGame {
         ResultView.printRacingResult(stringBuilder.toString());
     }
 
-    private static String positionsToString(Cars cars) {
+    private String positionsToString(Cars cars) {
         StringBuilder stringBuilder = new StringBuilder();
 
         for (Car car : cars.getList()) {
@@ -49,17 +37,12 @@ public class RacingGame {
         return stringBuilder.toString();
     }
 
-    private static Movable getMovable() {
+    private Movable getMovable() {
         return new Movable() {
             @Override
             public boolean isMovable() {
                 return new Random().nextInt(RANDOM_VALUE_MAX) >= MOVE_CONDITION;
             }
         };
-    }
-
-    public static void main(String[] args) {
-        gameSetting();
-        playGame();
     }
 }
